provide an upgrade path for slam/pcl from the SVN checkout to the archive
The CMake build system of PCL has the "feature" of automatically
listing everything that has a CMakeLists.txt. If, across the updates,
directories get deleted, this makes the build fail, which is what
happens if you update from the SVN to the tarball.

So, we must delete the SVN. Later on, we also will have to automatically
delete the current slam/pcl if we update the archive version.