Revert "make some packages next and stable"
This reverts commit 20cdcaca8ff42c4205ae680b5e649631b85a445d.

sonar_tritech never went to next. I just added it in a previous commit

auv_rel_pos_controller as well as auv_control should NOT go into next. We
discussed a general "AUV control" structure and IMO only that should go there.
Or, more precisely, we should discuss it on the ML.